app/requirements.txt
celery>=5,<6
channels
channels-redis
django>=4,<5
django-debug-toolbar>=4,<5
django-celery-beat
django-extensions
django-redis
emoji-country-flag
gunicorn
httpx
ipwhois
kaleido
mysqlclient
plotly
python-decouple
sentry-sdk
uvicorn[standard]
Werkzeug
